requirements.txt
Cython>=0.29.21
Kivy==2.0.0
paho-mqtt>=1.5.1
pyserial==3.5
